feat: add version controllers to routes and handle shutdown better
